fix(review): move computeGapHighFlags into the domain layer
The initial/correction gap split is a regulatory derivation, not a CSE
UI helper. Domain owns it so the CSE funnel and the XLSX export cannot
diverge.
