+++ packages/app/src/app/avis-cse/etape/[step]/page.tsx
@@ -3,13 +3,13 @@ import { campaignYearDimension, FunnelStepTracker } from "~/modules/analytics";
 import {
 	CSE_FUNNEL,
 	computeContentTypeColumns,
-	computeGapHighFlags,
 	mapOpinionsFromDb,
 	Step1Opinions,
 	Step2Upload,
 	TOTAL_STEPS,
 } from "~/modules/cseOpinion";
 import { getCseOpinionPreviousHref } from "~/modules/declaration-remuneration/shared/complianceNavigation";
+import { computeGapHighFlags } from "~/modules/domain";
 import { auth } from "~/server/auth";
 import { getCampaignDeadlines } from "~/server/db/getCampaignDeadlines";
 import { api } from "~/trpc/server";
