+++ packages/app/src/modules/cseOpinion/index.ts
@@ -2,7 +2,6 @@ export { ConfirmationPage } from "./ConfirmationPage";
 export { CseOpinionLayout } from "./CseOpinionLayout";
 export { computeContentTypeColumns } from "./contentTypeColumns";
 export { CSE_FUNNEL } from "./funnelConfig";
-export { computeGapHighFlags } from "./gapHighFlags";
 export { mapOpinionsFromDb } from "./mapOpinionsFromDb";
 export { Step1Opinions } from "./Step1Opinions";
 export { Step2Upload } from "./Step2Upload";
