+++ packages/app/src/modules/domain/index.ts
@@ -181,6 +181,7 @@ export type { GipGapReference } from "./shared/gap";
 export {
 	computeGap,
 	computeGapBetween,
+	computeGapHighFlags,
 	computeGapRatio,
 	computeTotal,
 	gapDirection,
